+++ danling/__init__.py
@@ -1,14 +1,14 @@
-from .logging import AverageMeter
+from .metrics import AverageMeter
 from .models import (FullyConnectedNetwork, MultiHeadAttention, SelfAttention, TransformerDecoder,
                      TransformerDecoderLayer, TransformerEncoder, TransformerEncoderLayer, UnitedPositionEmbedding)
 from .optim import LRScheduler
 from .registry import GlobalRegistry, Registry
-from .runner import BaseRunner, EpochRunner, StepRunner
+from .runner import EpochRunner, Runner, StepRunner
 from .tensors import NestedTensor
 from .utils import catch, load
 
 __all__ = [
-    "BaseRunner",
+    "Runner",
     "EpochRunner",
     "StepRunner",
     "Registry",
