+++ danling/registry/registry.py
@@ -1,46 +1,141 @@
+from collections.abc import Mapping
 from functools import wraps
 from typing import Callable, Optional
 
-from chanfig import NestedDict, OrderedDict
+from chanfig import FlatDict, NestedDict
 
 
 class Registry(NestedDict):
     """
-    Registry for components
+    Registry for components.
+
+    Registry inherits from NestedDict, this allows it has a NestedDict structure.
+
+    You could create a sub-registry by simply calling `registry.sub_registry = Registry`,
+    and access through `registry.sub_registry.register()`.
     """
 
-    def __init__(self, name):
+    override: bool = False
+
+    def __init__(self, override: bool = False):
         super().__init__()
-        self.setattr("__name__", name)
+        self.setattr("override", override)
 
-    def register(self, value: Optional[Callable] = None, name: Optional[str] = None):
-        # Register.register()
+    def register(self, component: Optional[Callable] = None, name: Optional[str] = None) -> Callable:
+        r"""
+        Register a new component
+
+        Parameters
+        ----------
+        component: Optional[Callable] = None
+            The component to register.
+        name: Optional[str] = component.__name__
+            The name of the component.
+
+        Returns
+        -------
+        component: Callable
+            The registered component.
+
+        Raises
+        ------
+        ValueError
+            If the component with the same name already exists and `Registry.override=False`.
+
+        Examples
+        --------
+        ```python
+        >>> registry = Registry("test")
+        >>> @registry.register
+        ... @registry.register("Module1")
+        ... class Module:
+        ...     def __init__(self, a, b):
+        ...         self.a = a
+        ...         self.b = b
+        >>> module = registry.register(Module, "Module2")
+        >>> registry
+        Registry(
+          (Module1): <class 'danling.registry.registry.Module'>
+          (Module): <class 'danling.registry.registry.Module'>
+          (Module2): <class 'danling.registry.registry.Module'>
+        )
+
+        ```
+        """
+
+        if isinstance(name, str) and name in self and not self.override:
+            raise ValueError(f"Component with name {name} already exists")
+
+        # Registry.register()
         if name is not None:
-            self.set(name, value)
+            self.set(name, component)
 
-        # @Register.register()
+        # @Registry.register()
         @wraps(self.register)
-        def register(value, name=None):
+        def register(component, name=None):
             if name is None:
-                name = value.__name__
-            self.set(name, value)
-            return value
+                name = component.__name__
+            self.set(name, component)
+            return component
 
-        # @Register.register
-        if callable(value) and name is None:
-            return register(value)
+        # @Registry.register
+        if callable(component) and name is None:
+            return register(component)
 
-        return lambda x: register(x, value)
+        return lambda x: register(x, component)
 
     lookup = NestedDict.get
 
     def build(self, name: str, *args, **kwargs):
-        if isinstance(name, OrderedDict) and not args and not kwargs:
-            name = name.pop("name")
-            kwargs = name
-        if not isinstance(name, str):
-            raise TypeError(f"name={name} should be a str, but got {type(name)}")
+        r"""
+        Build a component.
+
+        Parameters
+        ----------
+        name: str
+            The name of the component.
+        *args
+            The arguments to pass to the component.
+        **kwargs
+            The keyword arguments to pass to the component.
+
+        Returns
+        -------
+        component: Callable
+
+        Raises
+        ------
+        KeyError
+            If the component is not registered.
+
+        Examples
+        --------
+        ```python
+        >>> registry = Registry("test")
+        >>> @registry.register
+        ... class Module:
+        ...     def __init__(self, a, b):
+        ...         self.a = a
+        ...         self.b = b
+        >>> config = {"module": {"name": "Module", "a": 1, "b": 2}}
+        >>> # registry.register(Module)
+        >>> module = registry.build(config["module"])
+        >>> type(module)
+        <class 'danling.registry.registry.Module'>
+        >>> module.a
+        1
+        >>> module.b
+        2
+
+        ```
+        """
+
+        if isinstance(name, Mapping) and not args and not kwargs:
+            name, kwargs = name.pop("name"), name
         return self.get(name)(*args, **kwargs)
 
+    def __wrapped__(self):
+        pass
 
-GlobalRegistry = Registry("Global")
+
+GlobalRegistry = Registry()
