+++ danling/runner/__init__.py
@@ -1,10 +1,10 @@
-from .base_runner import BaseRunner
 from .epoch_runner import EpochRunner
+from .runner import Runner
 from .step_runner import StepRunner
 from .utils import ensure_dir, on_local_main_process, on_main_process
 
 __all__ = [
-    "BaseRunner",
+    "Runner",
     "EpochRunner",
     "StepRunner",
     "on_main_process",
