+++ danling/runner/base_runner.py
@@ -1,282 +1,414 @@
-import atexit
-import json
+from __future__ import annotations
+
 import logging
 import logging.config
 import os
-import random
-import shutil
 from collections.abc import Mapping
-from typing import Callable, Optional, Union
+from json import dumps as json_dumps
+from random import randint
+from typing import IO, TYPE_CHECKING, Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
 
-import numpy as np
 import torch
-from chanfig import Config, OrderedDict
-from torch import Tensor
-from torch import distributed as dist
-from torch.backends import cudnn
+from accelerate import Accelerator
+from chanfig import Config, FlatDict, NestedDict
+from chanfig.utils import JsonEncoder, YamlDumper
+from torch import nn, optim
+from yaml import dump as yaml_dump
 
-from danling.utils import catch
+from danling.utils import catch, is_json_serializable
 
-from .runner import Runner
-from .utils import on_main_process
+from .utils import ensure_dir
+
+if TYPE_CHECKING:
+    from torch.utils.tensorboard.writer import SummaryWriter
+
+PathStr = Union[os.PathLike, str, bytes]
+File = Union[PathStr, IO]
 
 
-class BaseRunner(Runner):
+class BaseRunner:
+
+    r"""
+    Base class for all runners.
+
+    Attributes
+    ----------
+    id: str = f"{self.name}-{self.seed}"
+    name: str = "danling"
+
+    seed: int = randint(0, 2 ** 32 - 1)
+    deterministic: bool = False
+
+    accelerator: Accelerator
+    accelerate: Dict[str, Any] = {}
+        keyword arguments for :class:`accelerate`.
+
+    steps: int = 0
+        Current running steps.
+    epochs: int = 0
+        Current running epochs.
+    step_begin: int = 0
+        End running steps.
+    epoch_begin: int = 0
+        End running epochs.
+    step_end: int = 0
+        End running steps.
+    epoch_end: int = 0
+        End running epochs.
+
+    model: Optional[nn.Module] = None
+    criterion: Optional[Tuple[nn.Module]] = None
+    optimizer: Optional[optim.Optimizer] = None
+    scheduler: Optional[optim.lr_scheduler._LRScheduler] = None
+
+    datasets: FlatDict
+        All datasets, should be in the form of ``{subset: dataset}``.
+    datasamplers: FlatDict
+        All datasamplers, should be in the form of ``{subset: datasampler}``.
+    dataloaders: FlatDict
+        All dataloaders, should be in the form of ``{subset: dataloader}``.
+
+    batch_size: int = 1
+
+    results: List[NestedDict] = []
+        All results, should be in the form of ``[{subset: {index: score}}]``.
+    index_set: Optional[str] = 'val'
+        The subset to calculate the core score.
+    index: str = 'loss'
+        The index to calculate the core score.
+
+    experiments_root: str = "experiments"
+        The root directory to save all experiments.
+    checkpoint_dir_name: str = "checkpoints"
+        The name of the directory under run_dir to save checkpoints.
+
+    log: bool = True
+        Whether to log the results.
+    logger: Optional[logging.Logger] = None
+    tensorboard: bool = False
+        Whether to use tensorboard.
+    writer: Optional[SummaryWriter] = None
     """
-    Set up everything for running a job
-    """
 
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
+    id: str = ""
+    name: str = "DanLing"
 
-        if self.seed is not None:
-            self.init_seed()
+    seed: int
+    deterministic: bool = False
 
-        if self.deterministic:
-            self.init_deterministic()
+    accelerator: Accelerator
+    accelerate: Dict[str, Any] = {}
 
-        if self.id is None:
+    steps: int = 0
+    epochs: int = 0
+    step_begin: int
+    epoch_begin: int
+    step_end: int
+    epoch_end: int
+    """step_begin, epoch_begin, step_end, epoch_end may not present in some runners, so they are not initialised."""
+
+    model: Optional[nn.Module] = None
+    criterion: Optional[Tuple[nn.Module]] = None
+    optimizer: Optional[optim.Optimizer] = None
+    scheduler: Optional[optim.lr_scheduler._LRScheduler] = None
+
+    datasets: FlatDict
+    datasamplers: FlatDict
+    dataloaders: FlatDict
+
+    batch_size: int = 1
+
+    results: List[NestedDict] = []
+    index_set: Optional[str] = None
+    index: str = "loss"
+
+    experiments_root: str = "experiments"
+    checkpoint_dir_name: str = "checkpoints"
+    log: bool = True
+    logger: Optional[logging.Logger] = None
+    tensorboard: bool = False
+    writer: Optional[SummaryWriter] = None
+
+    def __init__(self, *args, **kwargs):
+        super().__init__()
+        if len(args) == 1 and isinstance(args[0], FlatDict) and not kwargs:
+            args, kwargs = (), args[0]
+        self.__dict__ = NestedDict(*args, **kwargs)
+        if "seed" not in self:
+            self.seed = randint(0, 2**32 - 1)
+        if not self.id:
             self.id = f"{self.name}-{self.seed}"
+        self.accelerator = Accelerator(**self.accelerate)
+        self.datasets = FlatDict()
+        self.datasamplers = FlatDict()
+        self.dataloaders = FlatDict()
 
-        if self.log:
-            self.init_logger()
-            self.init_print()
-
-        if self.tensorboard:
-            self.writer = self.init_tensorboard()
-
-        atexit.register(self.print_result)
-
-    def init_distributed(self) -> None:
+    @property
+    def distributed(self):
         """
-        Set up distributed training
+        If runner is in distributed mode
         """
 
-        # pylint: disable=W0201
+        return self.num_processes > 1
 
-        dist.init_process_group(backend="nccl")
-        self.process_index = dist.get_rank()
-        self.num_processes = dist.get_world_size()
-        self.local_process_index = int(os.environ.get("LOCAL_RANK", -1))
-        torch.cuda.set_device(self.local_process_index)
-        self.is_main_process = self.process_index == 0
-        self.is_local_main_process = self.local_process_index == 0
-
-    def init_deterministic(self) -> None:
+    @property
+    def batch_size_actual(self) -> int:
         """
-        Set up deterministic
+        Actual batch size
         """
 
-        cudnn.benchmark = False
-        cudnn.deterministic = True
+        return self.batch_size * self.num_processes * getattr(self, "accum_steps", 1)
 
-    def init_seed(self) -> None:
+    @property
+    def best_result(self) -> Optional[NestedDict]:
         """
-        Set up random seed
-        """
-
-        if self.seed is None:
-            self.seed = random.randint(0, 100000)
-            if self.distributed:
-                self.seed = self.accelerator.gather(Tensor(self.seed).cuda()).unsqueeze(0).flatten()[0]
-        torch.manual_seed(self.seed + self.process_index)
-        torch.cuda.manual_seed(self.seed + self.process_index)
-        np.random.seed(self.seed + self.process_index)
-        random.seed(self.seed + self.process_index)
-
-    @on_main_process
-    def init_logger(self) -> None:
-        """
-        Set up logger
+        Best result
         """
 
-        # Why is setting up proper logging so !@?#! ugly?
-        logging.config.dictConfig(
-            {
-                "version": 1,
-                "disable_existing_loggers": False,
-                "formatters": {
-                    "standard": {"format": "%(asctime)s [%(levelname)s] %(name)s: %(message)s"},
-                },
-                "handlers": {
-                    "stdout": {
-                        "level": "INFO",
-                        "formatter": "standard",
-                        "class": "logging.StreamHandler",
-                        "stream": "ext://sys.stdout",
-                    },
-                    "logfile": {
-                        "level": "DEBUG",
-                        "formatter": "standard",
-                        "class": "logging.FileHandler",
-                        "filename": os.path.join(self.dir, "run.log"),
-                        "mode": "a",
-                    },
-                },
-                "loggers": {
-                    "": {
-                        "handlers": ["stdout", "logfile"],
-                        "level": "DEBUG",
-                        "propagate": True,
-                    },
-                },
-            }
-        )
-        logging.captureWarnings(True)
-        self.logger = logging.getLogger("runner")
-        self.logger.flush = lambda: [h.flush() for h in self.logger.handlers]
+        return self.results[-1 - self.scores[::-1].index(self.best_score)] if self.results else None
 
-    def init_print(self, process: Optional[int] = 0, precision: Optional[int] = 10) -> None:
+    @property
+    def latest_result(self) -> Optional[NestedDict]:
         """
-        Set up print
-        Only print from a specific process or force indicated
-        Replace default print function with logging.info
+        Latest Results
         """
 
-        torch.set_printoptions(precision=precision)
+        return self.results[-1] if self.results else None
 
-        logger = logging.getLogger("print")
-        logger.flush = lambda: [h.flush for h in logger.handlers]
-        import builtins as __builtin__  # pylint: disable=C0415
-
-        builtin_print = __builtin__.print
-
-        @catch
-        def print(*args, force=False, end="\n", file=None, flush=False, **kwargs):  # pylint: disable=W0622
-            if self.process_index == process or force:
-                if self.log:
-                    logger.info(*args, **kwargs)
-                else:
-                    builtin_print(*args, end=end, file=file, flush=flush, **kwargs)
-
-        __builtin__.print = print
-
-    @on_main_process
-    def init_tensorboard(self, *args, **kwargs):
-        from torch.utils.tensorboard.writer import SummaryWriter  # pylint: disable=C0415
-
-        return SummaryWriter(log_dir=self.dir, *args, **kwargs)
-
-    def init_lr(
-        self,
-        lr_scale_factor: Optional[float] = None,
-        batch_size_base: Optional[int] = None,
-    ) -> None:
+    @property
+    def scores(self) -> List[float]:
         """
-        Set up learning rate according to linear scaling rule
+        Scores
         """
 
-        # pylint: disable=W0201
+        if not self.results:
+            return []
+        index_set = self.index_set or next(reversed(self.results[-1]))
+        return [r[index_set][self.index] for r in self.results]
 
-        if lr_scale_factor is None:
-            if batch_size_base is None:
-                if batch_size_base := getattr(self, "batch_size_base", None) is None:
-                    raise ValueError("batch_size_base must be specified to auto scale lr")
-            lr_scale_factor = self.batch_size_actual / batch_size_base
-        self.lr_scale_factor = lr_scale_factor
-        self.lr = self.lr * self.lr_scale_factor
-        self.lr_final = self.lr_final * self.lr_scale_factor
+    @property
+    def best_score(self) -> Optional[float]:
+        """
+        Actual batch size
+        """
+
+        return self.best_fn(self.scores) if self.results else None
+
+    @property
+    def latest_score(self) -> Optional[float]:
+        """
+        Actual batch size
+        """
+
+        return self.scores[-1] if self.results else None
+
+    @staticmethod
+    def best_fn(scores: Sequence[float]) -> float:
+        """
+        Function to determine best score from a list of scores
+        """
+
+        return max(scores)
+
+    @property
+    def is_best(self) -> bool:
+        """
+        If current epoch is the best epoch
+        """
+
+        return self.latest_score == self.best_score
+
+    @property
+    def iters(self) -> int:
+        """
+        Number of iterations
+        """
+
+        return self.steps * self.batch_size_actual
+
+    @property
+    def progress(self) -> float:
+        """
+        Training Progress
+        """
+
+        if hasattr(self, "iter_end"):
+            return self.iters / self.iter_end
+        if hasattr(self, "step_end"):
+            return self.steps / self.step_end
+        if hasattr(self, "epoch_end"):
+            return self.epochs / self.epoch_end
+        return 0
+
+    @property
+    @ensure_dir
+    def dir(self) -> str:
+        """
+        Directory of experiment
+        """
+
+        return os.path.join(self.experiments_root, self.id)
+
+    @property
+    def log_path(self) -> str:
+        """
+        Path of log file
+        """
+
+        return os.path.join(self.dir, "run.log")
+
+    @property
+    @ensure_dir
+    def checkpoint_dir(self) -> str:
+        """
+        Directory of checkpoints
+        """
+
+        return os.path.join(self.dir, self.checkpoint_dir_name)
 
     @catch
-    @on_main_process
-    def save_checkpoint(self) -> None:
+    def save(self, obj: Any, f: File, on_main_process: bool = True) -> File:
         """
-        Save checkpoint to checkpoint_dir
+        Save object to a path or file
         """
-        latest_path = os.path.join(self.checkpoint_dir, "latest.pth")
-        self.save(self.state_dict(), latest_path)
-        if hasattr(self, "save_freq") and (self.epochs + 1) % self.save_freq == 0:
-            save_path = os.path.join(self.checkpoint_dir, f"epoch-{self.epochs}.pth")
-            shutil.copy(latest_path, save_path)
-        if self.is_best:
-            best_path = os.path.join(self.checkpoint_dir, "best.pth")
-            shutil.copy(latest_path, best_path)
 
-    def load_checkpoint(
-        self, checkpoint: Union[Mapping, str], override_config: bool = True, *args, **kwargs
-    ) -> None:  # pylint: disable=W1113
+        if on_main_process and self.is_main_process or not on_main_process:
+            self.accelerator.save(obj, f)
+        return f
+
+    @staticmethod
+    def load(f: File, *args, **kwargs) -> Any:
         """
-        Load runner from checkpoint
+        Load object from a path or file
         """
-        if isinstance(checkpoint, str):
-            if not os.path.isfile(checkpoint):
-                raise FileNotFoundError(f"checkpoint at {checkpoint} is not a file")
-            checkpoint = self.load(checkpoint, *args, **kwargs)
-        if override_config:
-            self.__dict__.update(checkpoint["runner"])
-        if self.model is not None and "model" in checkpoint:
-            self.model.load_state_dict(checkpoint["model"])
-        if self.optimizer is not None and "optimizer" in checkpoint:
-            self.optimizer.load_state_dict(checkpoint["optimizer"])
-        if self.scheduler is not None and "scheduler" in checkpoint:
-            self.scheduler.load_state_dict(checkpoint["scheduler"])
+
+        return torch.load(f, *args, **kwargs)
+
+    def __getattr__(self, name) -> Any:
+        if "accelerator" not in self:
+            raise RuntimeError(f"{self.__class__.__name__} is not properly initialised")
+        if hasattr(self.accelerator, name):
+            return getattr(self.accelerator, name)
+        raise AttributeError(f"{self.__class__.__name__} object has no attribute {name}")
+
+    def __contains__(self, name) -> bool:
+        return name in self.__dict__
+
+    def dict(self, cls: Callable = dict) -> Mapping:
+        """
+        Convert config to `cls`
+        """
+
+        # pylint: disable=C0103
+
+        ret = cls()
+        for k, v in self.__dict__.items():
+            if isinstance(v, FlatDict):
+                v = v.dict(cls)
+            if is_json_serializable(v):
+                ret[k] = v
+        return ret
+
+    @catch
+    def json(self, file: File, on_main_process: bool = True, *args, **kwargs) -> None:
+        """
+        Dump Runner to json file
+        """
+
+        if on_main_process and self.is_main_process or not on_main_process:
+            with FlatDict.open(file, mode="w") as fp:
+                fp.write(self.jsons(*args, **kwargs))
 
     @classmethod
-    def from_checkpoint(
-        cls, checkpoint, map_location: str = "cpu", convert_mapping: bool = True, *args, **kwargs
-    ) -> Runner:
-        """
-        Load Runner from checkpoint
+    def from_json(cls, file: File, *args, **kwargs) -> BaseRunner:
+        r"""
+        Construct Runner from json file.
+
+        This function calls `self.from_jsons()` to construct object from json string.
+        You may overwrite `from_jsons` in case something is not json serializable.
+
         """
 
-        checkpoint = cls.load(checkpoint, *args, map_location=map_location, **kwargs)
-        config = Config(**checkpoint["runner"]) if convert_mapping else checkpoint["runner"]
-        runner = cls(**config)
-        runner.load_checkpoint(checkpoint, override_config=False)
-        return runner
+        with FlatDict.open(file) as fp:
+            return cls.from_jsons(fp.read(), *args, **kwargs)
 
-    def state_dict(self, cls: Callable = dict) -> Mapping:
+    def jsons(self, *args, **kwargs) -> str:
         """
-        Return dict of all attributes for checkpoint
+        Dump Runner to json string
         """
-        if self.model is None:
-            raise ValueError("Model must be defined when calling state_dict")
-        model = self.accelerator.unwrap_model(self.model)
-        return cls(
-            runner=self.to(dict),
-            model=model.state_dict(),
-            optimizer=self.optimizer.state_dict() if self.optimizer else None,
-            scheduler=self.scheduler.state_dict() if self.scheduler else None,
-        )
+
+        if "cls" not in kwargs:
+            kwargs["cls"] = JsonEncoder
+        return json_dumps(self.dict(), *args, **kwargs)
+
+    @classmethod
+    def from_jsons(cls, string: str, *args, **kwargs) -> BaseRunner:
+        r"""
+        Construct Runner from json string.
+        """
+
+        return cls(**Config.from_jsons(string, *args, **kwargs))
 
     @catch
-    @on_main_process
-    def save_result(self) -> None:
+    def yaml(self, file: File, on_main_process: bool = True, *args, **kwargs) -> None:
         """
-        Save result to dir
+        Dump Runner to yaml file
         """
-        ret = {"id": self.id, "name": self.name}
-        result = self.latest_result
-        if isinstance(result, OrderedDict):
-            result = result.dict()
-        ret.update(result)  # This is slower but ensure id in the first
-        latest_path = os.path.join(self.dir, "latest.json")
-        with open(latest_path, "w") as f:
-            json.dump(ret, f, indent=4)
-        if self.is_best:
-            best_path = os.path.join(self.dir, "best.json")
-            shutil.copy(latest_path, best_path)
 
-    def print_result(self) -> None:
-        """
-        Print latest and best result
-        """
-        print(f"latest result: {self.latest_result}")
-        print(f"best result: {self.best_result}")
+        if on_main_process and self.is_main_process or not on_main_process:
+            with FlatDict.open(file, mode="w") as fp:
+                self.yamls(fp, *args, **kwargs)
 
-    def append_result(self, result) -> None:
-        """
-        Add latest result and update best result
-        """
-        self.results.append(result)
+    @classmethod
+    def from_yaml(cls, file: File, *args, **kwargs) -> BaseRunner:
+        r"""
+        Construct Runner from yaml file.
 
-    def step(self) -> None:
+        This function calls `self.from_yamls()` to construct object from yaml string.
+        You may overwrite `from_yamls` in case something is not yaml serializable.
         """
-        step optimizer and scheduler
+
+        with FlatDict.open(file) as fp:
+            return cls.from_yamls(fp.read(), *args, **kwargs)
+
+    def yamls(self, *args, **kwargs) -> str:
         """
-        if self.optimizer is not None:
-            self.optimizer.step()
-            self.optimizer.zero_grad()
-        if self.scheduler is not None:
-            self.scheduler.step()
-        self.steps += 1
+        Dump Runner to yaml string
+        """
+
+        if "Dumper" not in kwargs:
+            kwargs["Dumper"] = YamlDumper
+        return yaml_dump(self.dict(), *args, **kwargs)  # type: ignore
+
+    @classmethod
+    def from_yamls(cls, string: str, *args, **kwargs) -> BaseRunner:
+        r"""
+        Construct Runner from yaml string.
+        """
+
+        return cls(**Config.from_yamls(string, *args, **kwargs))
+
+    def _add_indent(self, s):
+        st = s.split("\n")
+        # don't do anything for single-line stuff
+        if len(st) == 1:
+            return s
+        first = st.pop(0)
+        st = [(2 * " ") + line for line in st]  # hardcode indent to 2
+        st = "\n".join(st)
+        st = first + "\n" + st
+        return st
+
+    def __repr__(self):
+        lines = []
+        for key, value in self.__dict__.items():
+            value_str = repr(value)
+            value_str = self._add_indent(value_str)
+            lines.append("(" + key + "): " + value_str)
+
+        main_str = self.__class__.__name__ + "("
+        if lines:
+            main_str += "\n  " + "\n  ".join(lines) + "\n"
+
+        main_str += ")"
+        return main_str
