+++ danling/runner/runner.py
@@ -1,349 +1,287 @@
-from __future__ import annotations
-
+import atexit
+import json
 import logging
 import logging.config
 import os
+import random
+import shutil
 from collections.abc import Mapping
-from typing import IO, TYPE_CHECKING, Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
+from typing import Callable, Optional, Union
 
+import numpy as np
 import torch
-from accelerate import Accelerator
-from chanfig import Config, NestedDict, OrderedDict
-from chanfig.config import JsonEncoder, YamlDumper, json_dumps, yaml_dump
-from torch import nn, optim
-from torch.utils import data
+from chanfig import Config, FlatDict
+from torch import Tensor
+from torch import distributed as dist
+from torch.backends import cudnn
 
-from danling.utils import catch, is_json_serializable
+from danling.utils import catch
 
-from .utils import ensure_dir
-
-if TYPE_CHECKING:
-    from torch.utils.tensorboard import SummaryWriter
-
-PathStr = Union[os.PathLike, str, bytes]
-File = Union[PathStr, IO]
+from .base_runner import BaseRunner
+from .utils import on_main_process
 
 
-class Runner:
+class Runner(BaseRunner):
+    """
+    Set up everything for running a job
+    """
 
-    id: str = None
-    name: str = "danling"
+    def __init__(self, *args, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
 
-    seed: int = 1031
-    deterministic: bool = False
+        if self.seed is not None:
+            self.init_seed()
 
-    experiment_dir: str = "experiments"
-    checkpoint_dir_name: str = "checkpoints"
+        if self.deterministic:
+            self.init_deterministic()
 
-    steps: int = 0
-    epochs: int = 0
+        if self.log:
+            self.init_logger()
+            self.init_print()
 
-    accelerator: Accelerator
-    accelerate: Dict[str, Any] = {}
+        if self.tensorboard:
+            self.writer = self.init_tensorboard()
 
-    model: Optional[nn.Module] = None
-    optimizer: Optional[optim.Optimizer] = None
-    scheduler: Optional[optim.lr_scheduler._LRScheduler] = None
+        atexit.register(self.print_result)
 
-    datasets: OrderedDict[str, data.Dataset] = OrderedDict()
-    datasamplers: OrderedDict[str, data.Sampler] = OrderedDict()
-    dataloaders: OrderedDict[str, data.DataLoader] = OrderedDict()
-
-    batch_size: int = 1
-
-    criterion: Tuple[nn.Module] = None
-
-    results: List[NestedDict[str, any]] = []
-    index_set: Optional[str] = None
-    index: str = "loss"
-
-    log: bool = True
-    logger: Optional[logging.Logger] = None
-    tensorboard: bool = False
-    writer: Optional[SummaryWriter] = None
-
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-        if len(args) == 1 and isinstance(args[0], OrderedDict) and not kwargs:
-            args, kwargs = (), args[0]
-        self.__dict__ = NestedDict()
-        self.__dict__.update(args)
-        self.__dict__.update(kwargs)
-        self.accelerator = Accelerator(**self.accelerate)
-
-    @property
-    def distributed(self):
+    def init_distributed(self) -> None:
         """
-        If runner is in distributed mode
+        Set up distributed training
         """
 
-        return self.num_processes > 1
+        # pylint: disable=W0201
 
-    @property
-    def batch_size_actual(self) -> int:
-        """
-        Actual batch size
-        """
+        dist.init_process_group(backend="nccl")
+        self.process_index = dist.get_rank()
+        self.num_processes = dist.get_world_size()
+        self.local_process_index = int(os.environ.get("LOCAL_RANK", -1))
+        torch.cuda.set_device(self.local_process_index)
+        self.is_main_process = self.process_index == 0
+        self.is_local_main_process = self.local_process_index == 0
 
-        return self.batch_size * self.num_processes * getattr(self, "accum_steps", 1)
-
-    @property
-    def best_result(self) -> NestedDict:
+    def init_deterministic(self) -> None:
         """
-        Best result
-        """
-
-        return self.results[-1 - self.scores[::-1].index(self.best_score)] if self.results else None
-
-    @property
-    def latest_result(self) -> NestedDict:
-        """
-        Latest Results
+        Set up deterministic
         """
 
-        return self.results[-1] if self.results else None
+        cudnn.benchmark = False
+        cudnn.deterministic = True
 
-    @property
-    def scores(self) -> List[float]:
+    def init_seed(self) -> None:
         """
-        Scores
-        """
-
-        if not self.results:
-            return []
-        index_set = self.index_set or next(reversed(self.results[-1]))
-        return [r[index_set][self.index] for r in self.results]
-
-    @property
-    def best_score(self) -> float:
-        """
-        Actual batch size
+        Set up random seed
         """
 
-        return self.best_fn(self.scores) if self.results else None
+        if self.seed is None:
+            self.seed = random.randint(0, 100000)
+            if self.distributed:
+                self.seed = self.accelerator.gather(Tensor(self.seed).cuda()).unsqueeze(0).flatten()[0]
+        torch.manual_seed(self.seed + self.process_index)
+        torch.cuda.manual_seed(self.seed + self.process_index)
+        np.random.seed(self.seed + self.process_index)
+        random.seed(self.seed + self.process_index)
 
-    @property
-    def latest_score(self) -> float:
+    @on_main_process
+    def init_logger(self) -> None:
         """
-        Actual batch size
-        """
-
-        return self.scores[-1] if self.results else None
-
-    @staticmethod
-    def best_fn(scores: Sequence[float]) -> float:
-        """
-        Function to determine best score from a list of scores
-        """
-
-        return max(scores)
-
-    @property
-    def is_best(self) -> bool:
-        """
-        If current epoch is the best epoch
+        Set up logger
         """
 
-        return self.latest_score == self.best_score
+        # Why is setting up proper logging so !@?#! ugly?
+        logging.config.dictConfig(
+            {
+                "version": 1,
+                "disable_existing_loggers": False,
+                "formatters": {
+                    "standard": {"format": "%(asctime)s [%(levelname)s] %(name)s: %(message)s"},
+                },
+                "handlers": {
+                    "stdout": {
+                        "level": "INFO",
+                        "formatter": "standard",
+                        "class": "logging.StreamHandler",
+                        "stream": "ext://sys.stdout",
+                    },
+                    "logfile": {
+                        "level": "DEBUG",
+                        "formatter": "standard",
+                        "class": "logging.FileHandler",
+                        "filename": self.log_path,
+                        "mode": "a",
+                    },
+                },
+                "loggers": {
+                    "": {
+                        "handlers": ["stdout", "logfile"],
+                        "level": "DEBUG",
+                        "propagate": True,
+                    },
+                },
+            }
+        )
+        logging.captureWarnings(True)
+        self.logger = logging.getLogger("runner")
+        self.logger.flush = lambda: [h.flush() for h in self.logger.handlers]
 
-    @property
-    def iters(self) -> int:
+    def init_print(self, process: Optional[int] = 0, precision: Optional[int] = 10) -> None:
         """
-        Number of iterations
+        Set up print.
+
+        Only print on a specific process or when force is indicated.
+
+        Replace default print function with logging.info.
         """
 
-        return self.steps * self.batch_size_actual
+        torch.set_printoptions(precision=precision)
 
-    @property
-    def progress(self) -> float:
+        logger = logging.getLogger("print")
+        logger.flush = lambda: [h.flush for h in logger.handlers]
+        import builtins as __builtin__  # pylint: disable=C0415
+
+        builtin_print = __builtin__.print
+
+        @catch
+        def print(*args, force=False, end="\n", file=None, flush=False, **kwargs):  # pylint: disable=W0622
+            if self.process_index == process or force:
+                if self.log:
+                    logger.info(*args, **kwargs)
+                else:
+                    builtin_print(*args, end=end, file=file, flush=flush, **kwargs)
+
+        __builtin__.print = print
+
+    @on_main_process
+    def init_tensorboard(self, *args, **kwargs):
+        from torch.utils.tensorboard.writer import SummaryWriter  # pylint: disable=C0415
+
+        return SummaryWriter(log_dir=self.dir, *args, **kwargs)
+
+    def init_lr(
+        self,
+        lr_scale_factor: Optional[float] = None,
+        batch_size_base: Optional[int] = None,
+    ) -> None:
         """
-        Training Progress
+        Set up learning rate according to linear scaling rule
         """
 
-        if hasattr(self, "iter_end"):
-            return self.iters / self.iter_end
-        if hasattr(self, "step_end"):
-            return self.steps / self.step_end
-        if hasattr(self, "epoch_end"):
-            return self.epochs / self.epoch_end
-        return 0
+        # pylint: disable=W0201
 
-    @property
-    @ensure_dir
-    def dir(self) -> str:
-        """
-        Directory of experiment
-        """
-
-        return os.path.join(self.experiment_dir, self.id)
-
-    @property
-    @ensure_dir
-    def checkpoint_dir(self) -> str:
-        """
-        Directory of checkpoints
-        """
-
-        return os.path.join(self.dir, self.checkpoint_dir_name)
+        if lr_scale_factor is None:
+            if batch_size_base is None:
+                if batch_size_base := getattr(self, "batch_size_base", None) is None:
+                    raise ValueError("batch_size_base must be specified to auto scale lr")
+            lr_scale_factor = self.batch_size_actual / batch_size_base
+        self.lr_scale_factor = lr_scale_factor
+        self.lr = self.lr * self.lr_scale_factor
+        self.lr_final = self.lr_final * self.lr_scale_factor
 
     @catch
-    def save(self, obj: Any, f: File, on_main_process: bool = True) -> None:
+    @on_main_process
+    def save_checkpoint(self) -> None:
         """
-        Save object to a path or file
+        Save checkpoint to checkpoint_dir
         """
+        latest_path = os.path.join(self.checkpoint_dir, "latest.pth")
+        self.save(self.state_dict(), latest_path)
+        if hasattr(self, "save_freq") and (self.epochs + 1) % self.save_freq == 0:
+            save_path = os.path.join(self.checkpoint_dir, f"epoch-{self.epochs}.pth")
+            shutil.copy(latest_path, save_path)
+        if self.is_best:
+            best_path = os.path.join(self.checkpoint_dir, "best.pth")
+            shutil.copy(latest_path, best_path)
 
-        if on_main_process and self.is_main_process or not on_main_process:
-            self.accelerator.save(obj, f)
-        return f
-
-    @staticmethod
-    def load(f: File, *args, **kwargs) -> Any:
+    def load_checkpoint(
+        self, checkpoint: Union[Mapping, str], override_config: bool = True, *args, **kwargs
+    ) -> None:  # pylint: disable=W1113
         """
-        Load object from a path or file
+        Load runner from checkpoint
         """
+        if isinstance(checkpoint, str):
+            if not os.path.isfile(checkpoint):
+                raise FileNotFoundError(f"checkpoint at {checkpoint} is not a file")
+            checkpoint = self.load(checkpoint, *args, **kwargs)
+        if override_config:
+            self.__dict__.update(checkpoint["runner"])
+        if self.model is not None and "model" in checkpoint:
+            self.model.load_state_dict(checkpoint["model"])
+        if self.optimizer is not None and "optimizer" in checkpoint:
+            self.optimizer.load_state_dict(checkpoint["optimizer"])
+        if self.scheduler is not None and "scheduler" in checkpoint:
+            self.scheduler.load_state_dict(checkpoint["scheduler"])
 
-        return torch.load(f, *args, **kwargs)
-
-    def __getattr__(self, name) -> Any:
-        if "accelerator" not in self:
-            raise RuntimeError(f"{self.__class__.__name__} is not properly initialised")
-        if hasattr(self.accelerator, name):
-            return getattr(self.accelerator, name)
-        raise AttributeError(f'"Runner" object has no attribute "{name}"')
-
-    def __contains__(self, name) -> bool:
-        return name in self.__dict__
-
-    def __repr__(self):
-        r"""
-        Representation of OrderedDict.
-
-        Example:
-        ```python
-        >>> d = OrderedDict(a=1, b=2, c=3)
-        >>> repr(d)
-        'OrderedDict(\n  (a): 1\n  (b): 2\n  (c): 3\n)'
-
-        ```
+    @classmethod
+    def from_checkpoint(
+        cls, checkpoint, map_location: str = "cpu", convert_mapping: bool = True, *args, **kwargs
+    ) -> BaseRunner:
         """
-
-        lines = []
-        for key, value in self.__dict__.items():
-            value_str = repr(value)
-            value_str = self._add_indent(value_str)
-            lines.append("(" + key + "): " + value_str)
-
-        main_str = self.__class__.__name__ + "("
-        if lines:
-            main_str += "\n  " + "\n  ".join(lines) + "\n"
-
-        main_str += ")"
-        return main_str
-
-    def _add_indent(self, s):
-        st = s.split("\n")
-        # don't do anything for single-line stuff
-        if len(st) == 1:
-            return s
-        first = st.pop(0)
-        st = [(2 * " ") + line for line in st]  # hardcode indent to 2
-        st = "\n".join(st)
-        st = first + "\n" + st
-        return st
-
-    def to(self, cls: Callable = dict) -> Mapping:  # pylint: disable=C0103
-        """
-        Convert config to `cls`
+        Load Runner from checkpoint.
         """
 
-        ret = cls()
-        for k, v in self.__dict__.items():
-            if isinstance(v, OrderedDict):
-                v = v.to(cls)
-            if is_json_serializable(v):
-                ret[k] = v
-        return ret
+        checkpoint = cls.load(checkpoint, *args, map_location=map_location, **kwargs)
+        config = Config(**checkpoint["runner"]) if convert_mapping else checkpoint["runner"]
+        runner = cls(**config)
+        runner.load_checkpoint(checkpoint, override_config=False)
+        return runner
 
-    convert = to
+    def state_dict(self, cls: Callable = dict) -> Mapping:
+        """
+        Return dict of all attributes for checkpoint.
+        """
+
+        if self.model is None:
+            raise ValueError("Model must be defined when calling state_dict")
+        model = self.accelerator.unwrap_model(self.model)
+        return cls(
+            runner=self.dict(),
+            model=model.state_dict(),
+            optimizer=self.optimizer.state_dict() if self.optimizer else None,
+            scheduler=self.scheduler.state_dict() if self.scheduler else None,
+        )
 
     @catch
-    def json(self, file: File, on_main_process: bool = True, *args, **kwargs) -> None:
-        """
-        Dump Runner to json file
-        """
-
-        if on_main_process and self.is_main_process or not on_main_process:
-            with NestedDict.open(file, mode="w") as fp:
-                fp.write(self.jsons(*args, **kwargs))
-
-    @classmethod
-    def from_json(cls, file: File, *args, **kwargs) -> Runner:
+    @on_main_process
+    def save_result(self) -> None:
         r"""
-        Construct Runner from json file.
-
-        This function calls `self.from_jsons()` to construct object from json string.
-        You may overwrite `from_jsons` in case something is not json serializable.
-
+        Save result to `runner.dir`.
         """
 
-        with open(file) as fp:
-            return cls.from_jsons(fp.read(), *args, **kwargs)
+        ret = {"id": self.id, "name": self.name}
+        result = self.latest_result
+        if isinstance(result, FlatDict):
+            result = result.dict()
+        ret.update(result)  # This is slower but ensure id in the first
+        latest_path = os.path.join(self.dir, "latest.json")
+        with FlatDict.open(latest_path, "w") as f:
+            json.dump(ret, f, indent=4)
+        if self.is_best:
+            best_path = os.path.join(self.dir, "best.json")
+            shutil.copy(latest_path, best_path)
 
-    def jsons(self, *args, **kwargs) -> str:
-        """
-        Dump Runner to json string
-        """
-
-        if "cls" not in kwargs:
-            kwargs["cls"] = JsonEncoder
-        return json_dumps(self.to(dict), *args, **kwargs)
-
-    @classmethod
-    def from_jsons(cls, string: str, *args, **kwargs) -> Runner:
+    def print_result(self) -> None:
         r"""
-        Construct Runner from json string.
-
+        Print latest and best result.
         """
 
-        return cls(**Config.from_jsons(string, *args, **kwargs))
+        print(f"latest result: {self.latest_result}")
+        print(f"best result: {self.best_result}")
 
-    @catch
-    def yaml(self, file: File, on_main_process: bool = True, *args, **kwargs) -> None:
-        """
-        Dump Runner to yaml file
-        """
-
-        if on_main_process and self.is_main_process or not on_main_process:
-            with NestedDict.open(file, mode="w") as fp:
-                self.yamls(fp, *args, **kwargs)
-
-    @classmethod
-    def from_yaml(cls, file: File, *args, **kwargs) -> Runner:
+    def append_result(self, result) -> None:
         r"""
-        Construct Runner from yaml file.
-
-        This function calls `self.from_yamls()` to construct object from yaml string.
-        You may overwrite `from_yamls` in case something is not yaml serializable.
-
+        Add latest result and update best result.
         """
 
-        with open(file) as fp:
-            return cls.from_yamls(fp.read(), *args, **kwargs)
+        self.results.append(result)
 
-    def yamls(self, *args, **kwargs) -> str:
-        """
-        Dump Runner to yaml string
-        """
-
-        if "Dumper" not in kwargs:
-            kwargs["Dumper"] = YamlDumper
-        return yaml_dump(self.to(dict), *args, **kwargs)  # type: ignore
-
-    @classmethod
-    def from_yamls(cls, string: str, *args, **kwargs) -> Runner:
+    def step(self, zero_grad: bool = True) -> None:
         r"""
-        Construct Runner from yaml string.
-
+        Step optimizer and scheduler.
         """
 
-        return cls(**Config.from_yamls(string, *args, **kwargs))
+        if self.optimizer is not None:
+            self.optimizer.step()
+            if zero_grad:
+                self.optimizer.zero_grad()
+        if self.scheduler is not None:
+            self.scheduler.step()
+        self.steps += 1
